+++ packages/core-js/actual/typed-array/index.js
@@ -2,6 +2,7 @@
 var parent = require('../../stable/typed-array');
 require('../../modules/esnext.uint8-array.from-base64');
 require('../../modules/esnext.uint8-array.from-hex');
+require('../../modules/esnext.uint8-array.set-from-base64');
 require('../../modules/esnext.uint8-array.set-from-hex');
 require('../../modules/esnext.uint8-array.to-base64');
 require('../../modules/esnext.uint8-array.to-hex');